web/requirements.txt
Django==2.0.10
django-debug-toolbar==1.11
djangorestframework==3.7.7
drf-writable-nested==0.5.1
gunicorn==19.7.1
mysqlclient==1.3.12
pytz==2018.9
sqlparse==0.3.0
